requirements.txt
absl-py==2.0.0
ale-py==0.8.1
appdirs==1.4.4
autopep8==2.0.4
autorom==0.4.2
autorom-accept-rom-license==0.6.1
box2d-py==2.3.5
cachetools==5.3.2
cloudpickle==3.0.0
crc32c==2.3.post0
decorator==4.4.2
docker-pycreds==0.4.0
docstring-parser==0.15
etils==1.6.0
ffmpeg==1.4
fsspec==2023.12.2
gitdb==4.0.11
gitpython==3.1.40
glfw==2.6.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
gymnasium==0.29.1
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
markdown==3.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
moviepy==1.0.3
mujoco==3.1.1
opencv-python==4.9.0.80
pandas==2.1.4
pip==23.3.2
proglog==0.1.10
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pygame==2.1.3
pyopengl==3.1.7
pytz==2023.3.post1
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
sentry-sdk==1.39.1
setproctitle==1.3.3
setuptools==69.0.2
shimmy==1.3.0
shtab==1.6.5
smmap==5.0.1
soundfile==0.12.1
stable-baselines3==2.2.1
swig==4.1.1.post1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
tomli==2.0.1
torch-tb-profiler==0.4.3
torchsummary==1.5.1
tqdm==4.66.1
tyro==0.6.4
tzdata==2023.3
vastai==0.2.0
wandb==0.16.2
wheel==0.42.0
zipp==3.17.0